Fix enum class usage in GUI visualizers, tests, and pyOpenMS
Update code to use proper enum class syntax after enum migration:
- GUI visualizers: Add static_cast for array indexing and setCurrentIndex
- C++ tests: Use fully qualified enum names (e.g., Class::EnumType::VALUE)
- pyOpenMS tests: Update enum references (Type_IonDetector -> Type,
  MZTrafoModel_MODELTYPE -> MZTrafoModel.MODELTYPE, etc.)

+++ src/openms_gui/source/VISUAL/SpectraIDViewTab.cpp
@@ -1051,7 +1051,7 @@ namespace OpenMS
     table_widget_->setAtBottomRow(spectrum.getRT(), Clmn::RT, background_color);
 
     // scan mode
-    table_widget_->setAtBottomRow(QString::fromStdString(spectrum.getInstrumentSettings().NamesOfScanMode[spectrum.getInstrumentSettings().getScanMode()]), Clmn::SCANTYPE, background_color);
+    table_widget_->setAtBottomRow(QString::fromStdString(spectrum.getInstrumentSettings().NamesOfScanMode[static_cast<size_t>(spectrum.getInstrumentSettings().getScanMode())]), Clmn::SCANTYPE, background_color);
 
     // zoom scan
     table_widget_->setAtBottomRow(spectrum.getInstrumentSettings().getZoomScan() ? "yes" : "no", Clmn::ZOOM, background_color);
